+++ android/app/src/main/java/com/ltonetwork/universal/MainActivity.java
@@ -34,20 +34,14 @@ public class MainActivity extends ReactActivity {
    * DefaultReactActivityDelegate} which allows you to easily enable Fabric and Concurrent React
    * (aka React 18) with two boolean flags.
    */
-
-    @Override
-    protected ReactActivityDelegate createReactActivityDelegate() {
-      return new ReactActivityDelegateWrapper(this,
-        new DefaultReactActivityDelegate(
-          this,
-          getMainComponentName(),
-          // If you opted-in for the New Architecture, we enable the Fabric Renderer.
-          DefaultNewArchitectureEntryPoint.getFabricEnabled(), // fabricEnabled
-          // If you opted-in for the New Architecture, we enable Concurrent React (i.e. React 18).
-          DefaultNewArchitectureEntryPoint.getConcurrentReactEnabled() // concurrentRootEnabled
-        )
-      );
-    }
+  @Override
+  protected ReactActivityDelegate createReactActivityDelegate() {
+    return new ReactActivityDelegateWrapper(this, BuildConfig.IS_NEW_ARCHITECTURE_ENABLED, new DefaultReactActivityDelegate(
+        this,
+        getMainComponentName(),
+        // If you opted-in for the New Architecture, we enable the Fabric Renderer.
+        DefaultNewArchitectureEntryPoint.getFabricEnabled()));
+  }
 
   /**
    * Align the back button behavior with Android S
