+++ android/app/src/main/res/values/strings.xml
@@ -1,7 +1,3 @@
 <resources>
-  <!-- Below line is handled by '@expo/configure-splash-screen' command and it's discouraged to modify it manually -->
-  <string name="app_name">LTOWallet</string>
-   <string name="expo_splash_screen_resize_mode" translatable="false">cover</string>
-  <string name="expo_splash_screen_status_bar_translucent" translatable="false">true</string>
-  <string name="expo_system_ui_user_interface_style" translatable="false">light</string>
+    <string name="app_name">LTOWallet</string>
 </resources>
