+++ android/gradle.properties
@@ -26,7 +26,7 @@ android.useAndroidX=true
 android.enableJetifier=true
 
 # Version of flipper SDK to use with React Native
-FLIPPER_VERSION=0.125.0
+FLIPPER_VERSION=0.182.0
 
 # Use this property to specify which architecture you want to build.
 # You can also override it from the CLI using
@@ -44,10 +44,6 @@ newArchEnabled=false
 # If set to false, you will be using JSC instead.
 hermesEnabled=true
 
-# The hosted JavaScript engine
-# Supported values: expo.jsEngine = "hermes" | "jsc"
-expo.jsEngine=hermes
-
 # Enable GIF support in React Native images (~200 B increase)
 expo.gif.enabled=true
 # Enable webp support in React Native images (~85 KB increase)
@@ -56,4 +52,5 @@ expo.webp.enabled=true
 # Disabled by default because iOS doesn't support animated webp
 expo.webp.animated=false
 
-android.disableAutomaticComponentCreation=true
+# Enable network inspector
+EX_DEV_CLIENT_NETWORK_INSPECTOR=true
