+++ package.json
@@ -2,7 +2,7 @@
   "name": "@ltonetwork/universal-wallet",
   "version": "0.0.1",
   "scripts": {
-    "start": "react-native start",
+    "start": "expo start --dev-client",
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "clean:android": "cd android && ./gradlew clean",
@@ -55,7 +55,7 @@
     "punycode": "^1.4.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.72.3",
+    "react-native": "0.72.4",
     "react-native-biometrics": "^3.0.1",
     "react-native-collapsible": "^1.6.0",
     "react-native-crypto": "^2.2.0",
